backend/requirements.txt
Flask==2.2.3          # Flask for backend
requests==2.31.0      # To make API requests to OpenWeatherMap
python-dotenv==1.0.0  # To manage environment variables
